+++ src/clint.cpp
@@ -121,25 +121,13 @@ static bool clint_peek(const pma_entry &pma, const machine &m, uint64_t page_off
             base(clint_mtimecmp_rel_addr) != base(clint_mtime_rel_addr) &&
             base(clint_mtime_rel_addr) != base(clint_msip0_rel_addr),
         "code expects msip0, mtimcmp, and mtime to be in different pages");
-    // There are 3 non-pristine pages: base(CLINT_MSIP0_REL_ADDR), base(CLINT_MTIMECMP_REL_ADDR), and
-    // base(CLINT_MTIME_REL_ADDR)
+    // There are 1 non-pristine page: base(CLINT_MTIMECMP_REL_ADDR)
+    // Both base(CLINT_MSIP0_REL_ADDR), base(CLINT_MTIME_REL_ADDR) contain only derived values, and therefore
+    // do not enter the Merkle tree. mtime is derived from mcycle, and msip0 is derived from mip
     switch (page_offset) {
-        case base(clint_msip0_rel_addr):
-            // This page contains only msip (which is either 0 or 1)
-            // Since we are little-endian, we can simply write the bytes
-            memset(scratch, 0, PMA_PAGE_SIZE);
-            aliased_aligned_write<uint64_t>(scratch + offset(clint_msip0_rel_addr),
-                (m.read_mip() & MIP_MSIP_MASK) == MIP_MSIP_MASK);
-            *page_data = scratch;
-            return true;
         case base(clint_mtimecmp_rel_addr):
             memset(scratch, 0, PMA_PAGE_SIZE);
             aliased_aligned_write<uint64_t>(scratch + offset(clint_mtimecmp_rel_addr), m.read_clint_mtimecmp());
-            *page_data = scratch;
-            return true;
-        case base(clint_mtime_rel_addr):
-            memset(scratch, 0, PMA_PAGE_SIZE);
-            aliased_aligned_write<uint64_t>(scratch + offset(clint_mtime_rel_addr), rtc_cycle_to_time(m.read_mcycle()));
             *page_data = scratch;
             return true;
         default: