+++ src/machine.cpp
@@ -1413,6 +1413,7 @@ void machine::dump_pmas(void) const {
         }
         std::array<char, 256> filename{};
         (void) sprintf(filename.data(), "%016" PRIx64 "--%016" PRIx64 ".bin", pma.get_start(), pma.get_length());
+        std::cerr << "writing to " << filename.data() << '\n';
         auto fp = unique_fopen(filename.data(), "wb");
         for (uint64_t page_start_in_range = 0; page_start_in_range < pma.get_length();
              page_start_in_range += PMA_PAGE_SIZE) {
@@ -1420,8 +1421,15 @@ void machine::dump_pmas(void) const {
             auto peek = pma.get_peek();
             if (!peek(pma, *this, page_start_in_range, &page_data, scratch.get())) {
                 throw std::runtime_error{"peek failed"};
-            } else if (page_data && fwrite(page_data, 1, PMA_PAGE_SIZE, fp.get()) != PMA_PAGE_SIZE) {
-                throw std::system_error{errno, std::generic_category(), "error writing to '"s + filename.data() + "'"s};
+            } else {
+                if (!page_data) {
+                    memset(scratch.get(), 0, PMA_PAGE_SIZE);
+                    page_data = scratch.get();
+                }
+                if (fwrite(page_data, 1, PMA_PAGE_SIZE, fp.get()) != PMA_PAGE_SIZE) {
+                    throw std::system_error{errno, std::generic_category(),
+                        "error writing to '"s + filename.data() + "'"s};
+                }
             }
         }
     }